\name{shinyAppFunctions-package}
\alias{shinyAppFunctions-package}
\alias{shinyAppFunctions}
\docType{package}
\title{
\packageTitle{shinyAppFunctions}
}
\description{
\packageDescription{shinyAppFunctions}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{shinyAppFunctions}
\packageIndices{shinyAppFunctions}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{shinyAppFunctions}

Maintainer: \packageMaintainer{shinyAppFunctions}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
